% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arpt.R
\name{arpt}
\alias{arpt}
\title{arpt}
\usage{
arpt(dataset, pz = 0.6, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is Z index
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

