% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dualelasticity.R
\name{dualelasticity}
\alias{dualelasticity}
\title{dualelasticity}
\usage{
dualelasticity(dataset, samp = 50)
}
\description{
This is elasticity function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

