% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualelasticity2.R
\name{dualelasticity2}
\alias{dualelasticity2}
\title{delasticity22}
\usage{
dualelasticity2(dataset)
}
\description{
This is elasticity function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

