% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualelasticity3.R
\name{dualelasticity3}
\alias{dualelasticity3}
\title{delasticity23}
\usage{
dualelasticity3(dataset)
}
\description{
This is elasticity function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

