% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{gini2}
\usage{
gini(dataset, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is gini function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

