\name{laeken2-package}
\alias{laeken2-package}
\alias{laeken2}
\docType{package}
\title{
The package calculates indexes to measure the risk of poverty and 
social exclusion. Test for TIPs and Lorenz curves are given.
}
\description{
More about what it does (maybe more than one line)
This package was made with project support of Junta de Andalucia
}
\details{
\tabular{ll}{
Package: \tab laeken2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-12\cr
License: \tab GPL 2\cr
}
An overview of how to use the package, including the most
important functions
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo 
Maintainer: Angel Berihuete <angel.berihuete@uca.es>
}
\references{
Some references to the theory 
}
Optionally other standard keywords, one per line, from file
KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
simple examples of the most important functions
}
