% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadLCS.R
\name{loadLCS}
\alias{loadLCS}
\title{loadLCS}
\usage{
loadLCS(lcs_d_file, lcs_h_file)
}
\description{
Function to load the LCS dataset from INE
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

