% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miuc.R
\name{miuc}
\alias{miuc}
\title{miuc}
\usage{
miuc(dataset, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is mean income person function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

