% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OmegaTIP.R
\name{OmegaTIP}
\alias{OmegaTIP}
\title{OmegaTIP}
\usage{
OmegaTIP(dataset, arpt.value, normalization = FALSE, samp)
}
\value{
ToDo
}
\description{
ToDo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

