% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsr.R
\name{qsr}
\alias{qsr}
\title{qsr2}
\usage{
qsr(dataset, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is quintile share ratio function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

