% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmpg.R
\name{rmpg}
\alias{rmpg}
\title{rmpg2}
\usage{
rmpg(dataset, arpt.value, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is the relative median at-risk-poverty gap function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

