% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s1.R
\name{s1}
\alias{s1}
\title{s1}
\usage{
s1(dataset, arpt.value, norm = FALSE, ci = FALSE, rep = 1000,
  verbose = FALSE)
}
\description{
This is maximun of TIP curve function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

