% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testGL.R
\name{testGL}
\alias{testGL}
\title{testGL}
\usage{
testGL(dataset1, dataset2, generalized = FALSE, samplesize = 10)
}
\value{
ToDo
}
\description{
ToDo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

