% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testTIP.R
\name{testTIP}
\alias{testTIP}
\title{test.TIP}
\usage{
testTIP(dataset1, dataset2, pz = 0.6, same.arpt.value = NULL,
  norm = FALSE, samplesize = 50)
}
\value{
ToDo
}
\description{
ToDo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

