% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testTIP2.R
\name{testTIP2}
\alias{testTIP2}
\title{test.TIP2}
\usage{
testTIP2(dataset1, dataset2, pz = 0.6, same.arpt.value = NULL,
  norm = FALSE, samplesize = 50)
}
\value{
ToDo
}
\description{
ToDo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

