% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arpr.R
\name{arpr}
\alias{arpr}
\title{arpr2}
\usage{
arpr(dataset, arpt.value, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is arpr2 function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

