% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arpt.R
\name{arpt}
\alias{arpt}
\title{People at risk of poverty or social exclusion}
\usage{
arpt(dataset, pz = 0.6, ci = FALSE, rep = 500, verbose = FALSE)
}
\arguments{
\item{pz}{A number between 0 and 1. The default is 0.6 representing
the 60 percent of the median equivalized disposable income.}

\item{ci}{logical; if  TRUE, 95% confidente interval is given
for the risk of poverty.}

\item{rep}{A number to do the confidence interval using boostrap
technique.}

\item{verbose}{logical; if TRUE the confindence interval is plotted.}
}
\description{
Estimate the threshold for the risk of poverty.
}
\details{
People are considered to be at risk of poverty 
when their income is less than a particular threshold. 
In the EU, the standard definition for that threshold has been 
set at 60 percent of the median equivalized disposable income.
}
\examples{
data(eusilc2)
ATdataset <- setupDataset(eusilc2, country = "AT")
arpt(ATdataset)

}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}
\references{
\url{http://ec.europa.eu/eurostat/statistics-explained/index.php/Glossary:At-risk-of-poverty_rate}
}

