\name{laeken2-package}
\alias{laeken2-package}
\alias{laeken2}
\docType{package}
\title{
What the package does (short line)
Something cool!
}
\description{
More about what it does (maybe more than one line)
This package was made with project support of Junta de Andalucia
}
\details{
\tabular{ll}{
Package: \tab laeken2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-12\cr
License: \tab What license is it under?\cr
}
An overview of how to use the package, including the most
important functions
}
\author{
Who wrote it

Maintainer: Who to complain to <angel.berihuete@uca.es>
Angel Berihuete is the maintainer of this package
}
\references{
Some references to the theory 
}
Optionally other standard keywords, one per line, from file
KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
simple examples of the most important functions
}
