% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmpg}
\alias{rmpg}
\title{rmpg2}
\usage{
rmpg(aux.data, z.index, ci = FALSE, rep = 1000, verbose = FALSE)
}
\description{
This is the relative median at-risk-poverty gap function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

