% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{s2}
\alias{s2}
\title{s22}
\usage{
s2(dataset, arpt.value, norm = FALSE, ci = FALSE, rep = 1000,
  verbose = FALSE)
}
\description{
This is s2 index function
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

