% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setupDataset}
\alias{setupDataset}
\title{setupDataset}
\usage{
setupDataset(dataset, s = "OCDE", country = "ES", region = "all",
  deflac = NULL, ppp.rates = NULL)
}
\description{
This is Z index
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

