% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupDataset.R
\name{setupDataset}
\alias{setupDataset}
\title{setupDataset}
\usage{
setupDataset(dataset, country = "ES", region = "all", s = "OCDE",
  deflac = NULL, ppp = FALSE)
}
\description{
This is Z index
}
\details{
Todo
}
\author{
A. Berihuete, C.D. Ramos and M.A. Sordo
}

